Move datetime functions to own module
